third_party/WebKit/LayoutTests/fast/mediastream/MediaStreamTrack-getSettings.html

<!DOCTYPE HTML>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

promise_test(function() {
  return navigator.mediaDevices.getUserMedia({audio: true})
      .then(function(s) {
    settings = s.getAudioTracks()[0].getSettings();
    assert_greater_than(Object.keys(settings).length, 0);
  });
}, 'An audio track returns settings');

promise_test(function() {
  return navigator.mediaDevices.getUserMedia({video: true})
      .then(function(s) {
    settings = s.getVideoTracks()[0].getSettings();
    assert_greater_than(Object.keys(settings).length, 0);
  });
}, 'A video track returns settings');

promise_test(function() {
  return navigator.mediaDevices.getUserMedia({video: true})
      .then(function(s) {
    settings = s.getVideoTracks()[0].getSettings();
    assert_greater_than(Object.keys(settings).length, 0);
    assert_true('frameRate' in settings,
                'Frame rate missing: ' + JSON.stringify(settings));
    assert_true('width' in settings,
                'Width missing: ' + JSON.stringify(settings));
    assert_true('height' in settings,
                'Height missing: ' + JSON.stringify(settings));
  });
}, 'A video track returns frame rate');

</script>
